refactor(system): 重构系统模块接口和数据结构
- 更新了多个控制器的接口设计，使用 CsPageArgs 封装分页参数
- 修改了部分 API 的 HTTP 方法，优化了请求结构
- 统一了查询接口的命名规范
- 调整了部分参数和返回值类型，提高了代码复用性

+++ cutejava/cutejava-common/src/main/java/cn/odboy/util/CsPageUtil.java
@@ -19,10 +19,10 @@ public final class CsPageUtil extends cn.hutool.core.util.PageUtil {
     /**
      * List 分页
      */
-    public static <T> List<T> softPaging(int page, int size, List<T> list) {
-        int pageIndex = page - 1;
-        int fromIndex = pageIndex * size;
-        int toIndex = pageIndex * size + size;
+    public static <T> List<T> softPaging(long page, long size, List<T> list) {
+        int pageIndex = Math.toIntExact(page - 1);
+        int fromIndex = Math.toIntExact(pageIndex * size);
+        int toIndex = Math.toIntExact(pageIndex * size + size);
         if (fromIndex > list.size()) {
             return Collections.emptyList();
         } else if (toIndex >= list.size()) {
